# docs/swagger/openapi/sird-openapi-dia-03.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "SIRD API",
                 "description":  "API Backend del Sistema Institucional de Repositorio Digital - SIRD",
                 "contact":  {
                                 "name":  "Gobierno Regional del Cusco",
                                 "email":  "soporte.sird@regioncusco.gob.pe"
                             },
                 "license":  {
                                 "name":  "Uso institucional"
                             },
                 "version":  "v1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
...
    "paths":  {
                  "/api/v1/admin/usuarios/{usuarioId}":  {
                                                             "get":  {
                                                                         "tags":  [
                                                                                      "M09 - Administración del Sistema"
                                                                                  ],
                                                                         "summary":  "Obtener usuario SIRD por ID",
                                                                         "operationId":  "obtenerPorId",
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "usuarioId",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "integer",
                                                                                                               "format":  "int64"
                                                                                                           }
                                                                                            }
                                                                                        ],
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "OK",
                                                                                                       "content":  {
                                                                                                                       "*/*":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/RespuestaApiUsuarioSirdResponse"
                                                                                                                                              }
                                                                                                                               }
                                                                                                                   }
                                                                                                   }
                                                                                       }
                                                                     },
                                                             "put":  {
                                                                         "tags":  [
                                                                                      "M09 - Administración del Sistema"
                                                                                  ],
                                                                         "summary":  "Actualizar usuario SIRD",
                                                                         "operationId":  "actualizar",
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "usuarioId",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "integer",
                                                                                                               "format":  "int64"
                                                                                                           }
                                                                                            }
                                                                                        ],
                                                                         "requestBody":  {
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ActualizarUsuarioSirdRequest"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         },
                                                                                             "required":  true
                                                                                         },
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "OK",
                                                                                                       "content":  {
                                                                                                                       "*/*":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/RespuestaApiUsuarioSirdResponse"
                                                                                                                                              }
                                                                                                                               }
                                                                                                                   }
                                                                                                   }
                                                                                       }
                                                                     }
                                                         },
                  "/api/v1/admin/feriados/{feriadoId}":  {
                                                             "put":  {
                                                                         "tags":  [
                                                                                      "M09 - Administración del Sistema"
                                                                                  ],
                                                                         "summary":  "Actualizar feriado institucional",
                                                                         "description":  "Actualiza los datos de un feriado institucional existente.",
                                                                         "operationId":  "actualizar_1",
                                                                         "parameters":  [
                                                                                            {
                                                                                                "name":  "feriadoId",
                                                                                                "in":  "path",
                                                                                                "required":  true,
                                                                                                "schema":  {
                                                                                                               "type":  "integer",
                                                                                                               "format":  "int64"
                                                                                                           }
                                                                                            }
                                                                                        ],
                                                                         "requestBody":  {
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ActualizarFeriadoInstitucionalRequest"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         },
                                                                                             "required":  true
                                                                                         },
                                                                         "responses":  {
                                                                                           "200":  {
                                                                                                       "description":  "OK",
                                                                                                       "content":  {
                                                                                                                       "*/*":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/RespuestaApiFeriadoInstitucionalResponse"
                                                                                                                                              }
                                                                                                                               }
                                                                                                                   }
                                                                                                   }
                                                                                       }
                                                                     }
                                                         },
...
                  "/api/v1/admin/configuracion/limites-ec/{tipoEspacio}":  {
                                                                               "put":  {
                                                                                           "tags":  [
                                                                                                        "M09 - Administración del Sistema"
                                                                                                    ],
                                                                                           "summary":  "Actualizar límite por tipo de Espacio Colaborativo",
                                                                                           "operationId":  "actualizar_2",
                                                                                           "parameters":  [
                                                                                                              {
                                                                                                                  "name":  "tipoEspacio",
                                                                                                                  "in":  "path",
                                                                                                                  "required":  true,
                                                                                                                  "schema":  {
                                                                                                                                 "type":  "string",
                                                                                                                                 "enum":  [
                                                                                                                                              "EXPEDIENTE",
                                                                                                                                              "PLAN",
                                                                                                                                              "PERFIL"
                                                                                                                                          ]
                                                                                                                             }
                                                                                                              }
                                                                                                          ],
                                                                                           "requestBody":  {
                                                                                                               "content":  {
                                                                                                                               "application/json":  {
                                                                                                                                                        "schema":  {
                                                                                                                                                                       "$ref":  "#/components/schemas/ActualizarLimiteEspacioColaborativoRequest"
                                                                                                                                                                   }
                                                                                                                                                    }
                                                                                                                           },
                                                                                                               "required":  true
                                                                                                           },
                                                                                           "responses":  {
                                                                                                             "200":  {
                                                                                                                         "description":  "OK",
                                                                                                                         "content":  {
                                                                                                                                         "*/*":  {
                                                                                                                                                     "schema":  {
                                                                                                                                                                    "$ref":  "#/components/schemas/RespuestaApiLimiteEspacioColaborativoResponse"
                                                                                                                                                                }
                                                                                                                                                 }
                                                                                                                                     }
                                                                                                                     }
                                                                                                         }
                                                                                       }
                                                                           },
...
                  "/api/v1/admin/usuarios":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "M09 - Administración del Sistema"
                                                                      ],
                                                             "summary":  "Listar usuarios SIRD",
                                                             "operationId":  "listar",
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "soloActivos",
                                                                                    "in":  "query",
                                                                                    "required":  false,
                                                                                    "schema":  {
                                                                                                   "type":  "boolean"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/RespuestaApiListUsuarioSirdResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         },
                                                 "post":  {
                                                              "tags":  [
                                                                           "M09 - Administración del Sistema"
                                                                       ],
                                                              "summary":  "Crear usuario SIRD",
                                                              "operationId":  "crear",
                                                              "requestBody":  {
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/CrearUsuarioSirdRequest"
                                                                                                                                      }
                                                                                                                       }
                                                                                              },
                                                                                  "required":  true
                                                                              },
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  {
                                                                                                            "*/*":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/RespuestaApiUsuarioSirdResponse"
                                                                                                                                   }
                                                                                                                    }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          }
                                             },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "RespuestaApiUsuarioSirdResponse":  {
                                                                               "type":  "object",
                                                                               "properties":  {
                                                                                                  "resultado":  {
                                                                                                                    "type":  "boolean"
                                                                                                                },
                                                                                                  "mensaje":  {
                                                                                                                  "type":  "string"
                                                                                                              },
                                                                                                  "datos":  {
                                                                                                                "$ref":  "#/components/schemas/UsuarioSirdResponse"
                                                                                                            },
                                                                                                  "fechaHora":  {
                                                                                                                    "type":  "string",
                                                                                                                    "format":  "date-time"
                                                                                                                }
                                                                                              }
                                                                           },
...
                                       "ActualizarFeriadoInstitucionalRequest":  {
                                                                                     "type":  "object",
                                                                                     "properties":  {
                                                                                                        "nombre":  {
                                                                                                                       "type":  "string",
                                                                                                                       "maxLength":  150,
                                                                                                                       "minLength":  3
                                                                                                                   },
                                                                                                        "fecha":  {
                                                                                                                      "type":  "string",
                                                                                                                      "format":  "date"
                                                                                                                  },
                                                                                                        "tipo":  {
                                                                                                                     "type":  "string",
                                                                                                                     "enum":  [
                                                                                                                                  "NACIONAL",
                                                                                                                                  "REGIONAL_CUSCO"
                                                                                                                              ]
                                                                                                                 },
                                                                                                        "recurrenteAnual":  {
                                                                                                                                "type":  "boolean"
                                                                                                                            }
                                                                                                    },
                                                                                     "required":  [
                                                                                                      "fecha",
                                                                                                      "nombre",
                                                                                                      "tipo"
                                                                                                  ]
                                                                                 },
...
                                   }
                   }
}